- Fixed a problem while building hexa1FE.cc on GCC 4.1 (CentOS 5.5). - Incorporated Stefan's fixes to OpenFOAM data reader due to negative   JacDets.

+++ source/Elements/3D/hexa1FE.cc
@@ -1684,7 +1684,7 @@ namespace CoupledField
          //perform damping
          while(l<30 && f_test >= f_old){
             l++;
-            Double dampFac = 1.0/pow(2.0,(Double)(l-1.0)); 
+            Double dampFac = 1.0/pow(2.0,(UInt)(l-1.0)); 
             xi_start = xi_k + (delta_xi * dampFac); 
             Local2GlobalCoord(f, xi_start, coordMat, NULL);
             f = f - globalPoint;
